xwayland: move monitor selection helper into xwayland namespace
